Update utility naming to helpers (aligns with SDK / other adapters).
