--- tests/Bot.Builder.Community.Adapters.Alexa.Tests/Utility/SkillRequestUtility.cs
@@ -1,15 +1,0 @@
-﻿using Alexa.NET.Request;
-using Newtonsoft.Json;
-
-namespace Bot.Builder.Community.Adapters.Alexa.Tests.Utility
-{
-    public static class SkillRequestUtility
-    {
-        public static SkillRequest CreateIntentRequest()
-        {
-            // This is a simple skill request originating from the Alexa test tool.
-            string message = "{\"version\":\"1.0\",\"session\":{\"new\":true,\"sessionId\":\"amzn1.echo-api.session.ac4e704a-8e17-4b31-aec0-21f49e45b47b\",\"application\":{\"applicationId\":\"amzn1.ask.skill.b5c74146-56d3-433d-a53d-6908f606da36\"},\"user\":{\"userId\":\"amzn1.ask.account.AFQLWI4PPKAJPA3OXHNVKFVNWW5VKRQZPPYKLOO5DSRNZXUNETLTN3L6U4DO2BSXLH725SFXWYMLRXXDLFE5VFT54JXALVOHKPL36P7RUBIFOCGLGOMDVW5VVTULAGLMVFSPY42OTRVDRRUGBRPQZOLS36QBNIYFCXFEJJNSRCAUQBMKNUTXWXTL3RM5ZOCTAYJKAYTCGJLYAPA\"}},\"context\":{\"System\":{\"application\":{\"applicationId\":\"amzn1.ask.skill.b5c74146-56d3-433d-a53d-6908f606da36\"},\"user\":{\"userId\":\"amzn1.ask.account.AFQLWI4PPKAJPA3OXHNVKFVNWW5VKRQZPPYKLOO5DSRNZXUNETLTN3L6U4DO2BSXLH725SFXWYMLRXXDLFE5VFT54JXALVOHKPL36P7RUBIFOCGLGOMDVW5VVTULAGLMVFSPY42OTRVDRRUGBRPQZOLS36QBNIYFCXFEJJNSRCAUQBMKNUTXWXTL3RM5ZOCTAYJKAYTCGJLYAPA\"},\"device\":{\"deviceId\":\"amzn1.ask.device.AFV6YMAWH3O3BGHJ2KHJTHFVMYNYMYNXAJYORR42KQ2R5A6VZOI7E4Z7ORW5GETGDTJKIXMFYY7IVBSYLQFACFAWROHRGQZA4CKAYH566GRXIAG3G2T4NPLLUFL3UYYYSZZGC44B4ZHDBVIA4N3EMAJULOZV7SUV6XW6KKYTXPBGLV6PB4WCG\",\"supportedInterfaces\":{}},\"apiEndpoint\":\"https://api.amazonalexa.com\",\"apiAccessToken\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImtpZCI6IjEifQ.eyJhdWQiOiJodHRwczovL2FwaS5hbWF6b25hbGV4YS5jb20iLCJpc3MiOiJBbGV4YVNraWxsS2l0Iiwic3ViIjoiYW16bjEuYXNrLnNraWxsLmI1Yzc0MTQ2LTU2ZDMtNDMzZC1hNTNkLTY5MDhmNjA2ZGEzNiIsImV4cCI6MTU4MzM2NzM2OCwiaWF0IjoxNTgzMzY3MDY4LCJuYmYiOjE1ODMzNjcwNjgsInByaXZhdGVDbGFpbXMiOnsiY29udGV4dCI6IkFBQUFBQUFBQUFBZkFsVitYOExOTERyQTZZeDgwY0VvS3dFQUFBQUFBQUQzQ3BkSktSblpSNXBiS2I1S3RnbTh3d2U5K2ROaDJTMjREbEhJVktaVGN4ODNvazhVS2lyREEvaUlEYlJHYmNXNHVaaU41amVsdFFqRnlPM2JGL0tVSmpxN3lDT2pQTVZQRElYdEhhZTk4UnVxZVR0b2ZQSFYxQ3RzSWVCaTdNSEM4THgvQmtuSlJydGt4SzJxbmlQa2xOT1pTOVhQS2crKzZkODhZc0dqaGRzVlVBT0FzVUJHZk44WURWTlF0NzFQZ0Q0WXNPWlNobXVzYmRsdGVDZjYwSUovVEhRN2RLb3RYMjhKOUZ0Z2FINXdYN0JsdjRiVnYwYVJ6ZDVzYUFFQlRLV2p3eVdmMHU2eGx3WVNreE13bTRDblNjNkZ1TDV3MW5KeTBiT0FVdW9OUnZhalgvdnFWSUtvVitxNis0TG11QjcyZHR0aDJmMDlzZWlOV3o1U2djTmxaUi9QYXhpM0RZR1pnRlZQcWs1S2xBMWNWNExPU0h1WXNLK2VaOEFPOVBQNE5UR3o1aXNMbHc9PSIsImNvbnNlbnRUb2tlbiI6bnVsbCwiZGV2aWNlSWQiOiJhbXpuMS5hc2suZGV2aWNlLkFGVjZZTUFXSDNPM0JHSEoyS0hKVEhGVk1ZTllNWU5YQUpZT1JSNDJLUTJSNUE2VlpPSTdFNFo3T1JXNUdFVEdEVEpLSVhNRllZN0lWQlNZTFFGQUNGQVdST0hSR1FaQTRDS0FZSDU2NkdSWElBRzNHMlQ0TlBMTFVGTDNVWVlZU1paR0M0NEI0WkhEQlZJQTROM0VNQUpVTE9aVjdTVVY2WFc2S0tZVFhQQkdMVjZQQjRXQ0ciLCJ1c2VySWQiOiJhbXpuMS5hc2suYWNjb3VudC5BRlFMV0k0UFBLQUpQQTNPWEhOVktGVk5XVzVWS1JRWlBQWUtMT081RFNSTlpYVU5FVExUTjNMNlU0RE8yQlNYTEg3MjVTRlhXWU1MUlhYRExGRTVWRlQ1NEpYQUxWT0hLUEwzNlA3UlVCSUZPQ0dMR09NRFZXNVZWVFVMQUdMTVZGU1BZNDJPVFJWRFJSVUdCUlBRWk9MUzM2UUJOSVlGQ1hGRUpKTlNSQ0FVUUJNS05VVFhXWFRMM1JNNVpPQ1RBWUpLQVlUQ0dKTFlBUEEifX0.Ve60KPtowEspIwMMoM0dNy2O7unALQQ0n6TBFnIGKqvYnTZe0VuJjlxhfBhzQrvhRcXy7tjiQI0lOlESQsomZh9pKSbdce2MkXIkvcGlc0sbPHdpAuh2bNMc3ZNGT8PaLcDzJpMojbzCgV5ZXZD7h3CYlb1zAaUSRM76Zi_MUgIv1IfZ9lL1c-Xh6cTo6tJI-KPso6pGeGtAIA3WenHxPuz0Dkiv_c_K0_ApJWN12MElfSu05rgv0r7ircO_LhQwbBpCRw7G3C7gm2WdCnEDso6FQ3BXyG_fRIczZReldqmn8W3agpL_xkkt3lB1m-2NPqbeC0JkKx7DPHywhfwS4Q\"},\"Viewport\":{\"experiences\":[{\"arcMinuteWidth\":246,\"arcMinuteHeight\":144,\"canRotate\":false,\"canResize\":false}],\"shape\":\"RECTANGLE\",\"pixelWidth\":1024,\"pixelHeight\":600,\"dpi\":160,\"currentPixelWidth\":1024,\"currentPixelHeight\":600,\"touch\":[\"SINGLE\"],\"video\":{\"codecs\":[\"H_264_42\",\"H_264_41\"]}},\"Viewports\":[{\"type\":\"APL\",\"id\":\"main\",\"shape\":\"RECTANGLE\",\"dpi\":160,\"presentationType\":\"STANDARD\",\"canRotate\":false,\"configuration\":{\"current\":{\"video\":{\"codecs\":[\"H_264_42\",\"H_264_41\"]},\"size\":{\"type\":\"DISCRETE\",\"pixelWidth\":1024,\"pixelHeight\":600}}}}]},\"request\":{\"type\":\"IntentRequest\",\"requestId\":\"amzn1.echo-api.request.8cccb332-ed63-4636-8915-caba1c0868b0\",\"timestamp\":\"2020-03-05T00:11:08Z\",\"locale\":\"en-US\",\"intent\":{\"name\":\"GetUserIntent\",\"confirmationStatus\":\"NONE\",\"slots\":{\"phrase\":{\"name\":\"phrase\",\"value\":\"fly\",\"resolutions\":{\"resolutionsPerAuthority\":[{\"authority\":\"amzn1.er-authority.echo-sdk.amzn1.ask.skill.b5c74146-56d3-433d-a53d-6908f606da36.phrase\",\"status\":{\"code\":\"ER_SUCCESS_NO_MATCH\"}}]},\"confirmationStatus\":\"NONE\",\"source\":\"USER\"}}}}}";
-            return JsonConvert.DeserializeObject<SkillRequest>(message);
-        }
-    }
-}
